Add Vertx.io based API with Swagger UI
This task attempts to tackle the task of introducing an alternative and
modern Vertex.io based API parallel to the existing servlet one and
based on the existing codebase. The new API will be fully compatible with
the currently existing one to ease migration. It is basically just the same
API running on Vertx and has a Swagger UI and the existing API documented
with Swagger annotations